+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @react-native-windows/automation-channel
 
-This log was last generated on Thu, 21 Sep 2023 05:13:50 GMT and should not be manually modified.
+This log was last generated on Fri, 22 Sep 2023 05:13:33 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.12.47
+
+Fri, 22 Sep 2023 05:13:33 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.708
 
 ## 0.12.46